refactor: AI => Separete nn classes into files
